requirements.txt
boto3
awscli
streamlit
langchain
pypdf
faiss-cpu
langchain_community
python-dotenv